dvf/requirements.txt
psycopg2-binary==2.9.9
flask==3.0.2
flask_sqlalchemy==3.1.1
gunicorn==21.2.0